+++ CHANGES.rst
@@ -9,6 +9,9 @@ Changelog
 - Update client to dynamically generate exception classes with special base
   types.
 
+- Add special handling for ``TypeError``, allowing us to eliminate the awkward
+  ``Py2Iterator`` wrapper.
+
 - Test/build script improvements
     - Enable pytest-xdist
     - Docs support